refactor(website): tickets-db.ts split (1096→<200 Re-Export-Index + 4 Sibling-Module) [T001155]
